Factor out post visitCompilationUnit visitor
